Reduce false positive detections with two frame prediction consensus
Occasionally, one frame of a motion event will be incorrectly predicted
to be some other cat, even if the adjacent frames are all of one cat.
The model is not perfect, but we the server can mitigate this false
positive by waiting for two frames to have matching predictions (and one
of them must be high confidence).